place function for lattice parser in own file
will need it's own file to handle all the capabilities
