+++ src/python/impactx/dashboard/Toolbar/file_imports/python/__init__.py
@@ -1,0 +1,6 @@
+from .lattice_helper import DashboardLatticeConfigParser as LatticeParser
+
+
+__all__ = [
+    "LatticeParser",
+]
